requirements.txt
streamlit
transformers
torch
tensorflow
tf-keras
pillow
